+++ macros/src/ldtk_entity.rs
@@ -7,6 +7,7 @@ static GRID_COORDS_ATTRIBUTE_NAME: &str = "grid_coords";
 static LDTK_ENTITY_ATTRIBUTE_NAME: &str = "ldtk_entity";
 static FROM_ENTITY_INSTANCE_ATTRIBUTE_NAME: &str = "from_entity_instance";
 static WITH_ATTRIBUTE_NAME: &str = "with";
+static DEFAULT_ATTRIBUTE_NAME: &str = "default";
 
 pub fn expand_ldtk_entity_derive(ast: syn::DeriveInput) -> proc_macro::TokenStream {
     let struct_name = &ast.ident;
@@ -96,6 +97,15 @@ pub fn expand_ldtk_entity_derive(ast: syn::DeriveInput) -> proc_macro::TokenStre
             field_constructions.push(expand_with_attribute(attribute, field_name, field_type));
             continue;
         }
+
+        let default = field
+            .attrs
+            .iter()
+            .find(|a| *a.path.get_ident().as_ref().unwrap() == DEFAULT_ATTRIBUTE_NAME);
+        if let Some(attribute) = default {
+            field_constructions.push(expand_default_attribute(attribute, field_name, field_type));
+            continue;
+        }
     }
 
     let generics = &ast.generics;
@@ -115,7 +125,7 @@ pub fn expand_ldtk_entity_derive(ast: syn::DeriveInput) -> proc_macro::TokenStre
                 tileset: Option<&bevy::prelude::Handle<bevy::prelude::Image>>,
                 tileset_definition: Option<&bevy_ecs_ldtk::prelude::TilesetDefinition>,
                 asset_server: &bevy::prelude::AssetServer,
-                texture_atlases: &mut bevy::prelude::Assets<bevy::prelude::TextureAtlas>,
+                texture_atlases: &mut bevy::prelude::Assets<bevy::prelude::TextureAtlasLayout>,
             ) -> Self {
                 Self {
                     #(#field_constructions)*
@@ -181,12 +191,12 @@ fn expand_sprite_sheet_bundle_attribute(
     match field_type {
         syn::Type::Path(syn::TypePath { path: syn::Path { segments, .. }, .. }) => {
             if let Some(last) = segments.last() {
-                if last.ident != *"SpriteSheetBundle" {
-                    panic!("#[sprite_sheet_bundle...] attribute should apply to a field of type bevy::prelude::SpriteSheetBundle")
+                if last.ident != *"LdtkSpriteSheetBundle" {
+                    panic!("#[sprite_sheet_bundle...] attribute should apply to a field of type bevy_ecs_ldtk::prelude::LdtkSpriteSheetBundle")
                 }
             }
         },
-        _ => panic!("#[sprite_sheet_bundle...] attribute should apply to a field of type bevy::prelude::SpriteSheetBundle")
+        _ => panic!("#[sprite_sheet_bundle...] attribute should apply to a field of type bevy_ecs_ldtk::prelude::LdtkSpriteSheetBundle")
     }
 
     match attribute
@@ -201,28 +211,28 @@ fn expand_sprite_sheet_bundle_attribute(
                 _ => panic!("First argument of #[sprite_sheet_bundle(...)] should be a string")
             };
             let tile_width = match nested_iter.next() {
-                Some(syn::NestedMeta::Lit(syn::Lit::Float(asset))) => asset.base10_parse::<f32>().unwrap(),
-                _ => panic!("Second argument of #[sprite_sheet_bundle(...)] should be a float")
+                Some(syn::NestedMeta::Lit(syn::Lit::Int(asset))) => asset.base10_parse::<u32>().unwrap(),
+                _ => panic!("Second argument of #[sprite_sheet_bundle(...)] should be an int")
             };
             let tile_height = match nested_iter.next() {
-                Some(syn::NestedMeta::Lit(syn::Lit::Float(asset))) => asset.base10_parse::<f32>().unwrap(),
-                _ => panic!("Third argument of #[sprite_sheet_bundle(...)] should be a float")
+                Some(syn::NestedMeta::Lit(syn::Lit::Int(asset))) => asset.base10_parse::<u32>().unwrap(),
+                _ => panic!("Third argument of #[sprite_sheet_bundle(...)] should be an int")
             };
             let columns = match nested_iter.next() {
-                Some(syn::NestedMeta::Lit(syn::Lit::Int(asset))) => asset.base10_parse::<usize>().unwrap(),
+                Some(syn::NestedMeta::Lit(syn::Lit::Int(asset))) => asset.base10_parse::<u32>().unwrap(),
                 _ => panic!("Fourth argument of #[sprite_sheet_bundle(...)] should be an int")
             };
             let rows = match nested_iter.next() {
-                Some(syn::NestedMeta::Lit(syn::Lit::Int(asset))) => asset.base10_parse::<usize>().unwrap(),
+                Some(syn::NestedMeta::Lit(syn::Lit::Int(asset))) => asset.base10_parse::<u32>().unwrap(),
                 _ => panic!("Fifth argument of #[sprite_sheet_bundle(...)] should be an int")
             };
             let padding = match nested_iter.next() {
-                Some(syn::NestedMeta::Lit(syn::Lit::Float(asset))) => asset.base10_parse::<f32>().unwrap(),
-                _ => panic!("Sixth argument of #[sprite_sheet_bundle(...)] should be a float")
+                Some(syn::NestedMeta::Lit(syn::Lit::Int(asset))) => asset.base10_parse::<u32>().unwrap(),
+                _ => panic!("Sixth argument of #[sprite_sheet_bundle(...)] should be an int")
             };
             let offset = match nested_iter.next() {
-                Some(syn::NestedMeta::Lit(syn::Lit::Float(asset))) => asset.base10_parse::<f32>().unwrap(),
-                _ => panic!("Seventh argument of #[sprite_sheet_bundle(...)] should be a float")
+                Some(syn::NestedMeta::Lit(syn::Lit::Int(asset))) => asset.base10_parse::<u32>().unwrap(),
+                _ => panic!("Seventh argument of #[sprite_sheet_bundle(...)] should be an int")
             };
             let index = match nested_iter.next() {
                 Some(syn::NestedMeta::Lit(syn::Lit::Int(asset))) => asset.base10_parse::<usize>().unwrap(),
@@ -230,20 +240,20 @@ fn expand_sprite_sheet_bundle_attribute(
             };
 
             quote! {
-                #field_name: bevy::prelude::SpriteSheetBundle {
-                    texture_atlas: texture_atlases.add(
-                        bevy::prelude::TextureAtlas::from_grid(
-                            asset_server.load(#asset_path).into(),
-                            bevy::prelude::Vec2::new(#tile_width, #tile_height),
-                            #columns, #rows, Some(bevy::prelude::Vec2::splat(#padding)),
-                            Some(bevy::prelude::Vec2::splat(#offset)),
-                        )
-                    ),
-                    sprite: bevy::prelude::TextureAtlasSprite {
-                        index: #index,
+                #field_name: LdtkSpriteSheetBundle{
+                        sprite_bundle: bevy::prelude::SpriteBundle {
+                        texture: asset_server.load(#asset_path).into(),
                         ..Default::default()
                     },
-                    ..Default::default()
+                    texture_atlas: bevy::prelude::TextureAtlas {
+                        layout: texture_atlases.add(
+                            bevy::prelude::TextureAtlasLayout::from_grid(
+                                bevy::prelude::UVec2::new(#tile_width, #tile_height),
+                                #columns, #rows, Some(bevy::prelude::UVec2::splat(#padding)),
+                                Some(bevy::prelude::UVec2::splat(#offset)),
+                            )),
+                        index: #index
+                    }
                 },
             }
         },
@@ -364,5 +374,23 @@ fn expand_with_attribute(
         _ => {
             panic!("#[with...] attribute should take the form #[with(function_name)]")
         }
+    }
+}
+
+fn expand_default_attribute(
+    attribute: &syn::Attribute,
+    field_name: &syn::Ident,
+    _: &syn::Type,
+) -> proc_macro2::TokenStream {
+    match attribute
+        .parse_meta()
+        .expect("Cannot parse #[default] attribute")
+    {
+        syn::Meta::Path(_) => {
+            quote! {
+                #field_name: Default::default(),
+            }
+        }
+        _ => panic!("#[default] attribute should take the form #[default]"),
     }
 }