+++ src/app/ldtk_int_cell.rs
@@ -154,6 +154,34 @@ use std::{collections::HashMap, marker::PhantomData};
 ///     damage: Damage,
 /// }
 /// ```
+///
+/// ### `#[default]`
+///
+/// Indicates that this component or bundle should be initialized using
+/// [`Default::default`].
+/// This can be useful when implementing `Default` for the whole `IntGridCell` is
+/// not easily possible, because some of the fields do not implement `Default`.
+///
+/// ```
+/// # use bevy::prelude::*;
+/// # use bevy_ecs_ldtk::prelude::*;
+/// # mod other_crate {
+/// #     #[derive(bevy::prelude::Component)]
+/// #     pub struct ForeignComponentWithNoDefault;
+/// # }
+/// # fn custom_constructor(_: IntGridCell) -> ForeignComponentWithNoDefault { todo!(); }
+/// # #[derive(Component, Default)]
+/// # struct Damage;
+/// use other_crate::ForeignComponentWithNoDefault;
+///
+/// #[derive(Bundle, LdtkIntCell)]
+/// pub struct MyBundle {
+///     #[default]
+///     damage: Damage,
+///     #[with(custom_constructor)]
+///     foreign: ForeignComponentWithNoDefault,
+/// }
+/// ```
 pub trait LdtkIntCell {
     /// The constructor used by the plugin when spawning additional components on IntGrid tiles.
     /// If you need access to more of the [World](bevy::prelude::World), you can create a system that queries for
@@ -191,21 +219,21 @@ impl<B: LdtkIntCell + Bundle> PhantomLdtkIntCell<B> {
 }
 
 pub trait PhantomLdtkIntCellTrait {
-    fn evaluate<'w, 's, 'a, 'b>(
+    fn evaluate<'a, 'b>(
         &self,
-        entity_commands: &'b mut EntityCommands<'w, 's, 'a>,
+        entity_commands: &'b mut EntityCommands<'a>,
         int_grid_cell: IntGridCell,
         layer_instance: &LayerInstance,
-    ) -> &'b mut EntityCommands<'w, 's, 'a>;
+    ) -> &'b mut EntityCommands<'a>;
 }
 
 impl<B: LdtkIntCell + Bundle> PhantomLdtkIntCellTrait for PhantomLdtkIntCell<B> {
-    fn evaluate<'w, 's, 'a, 'b>(
+    fn evaluate<'a, 'b>(
         &self,
-        entity_commands: &'b mut EntityCommands<'w, 's, 'a>,
+        entity_commands: &'b mut EntityCommands<'a>,
         int_grid_cell: IntGridCell,
         layer_instance: &LayerInstance,
-    ) -> &'b mut EntityCommands<'w, 's, 'a> {
+    ) -> &'b mut EntityCommands<'a> {
         entity_commands.insert(B::bundle_int_cell(int_grid_cell, layer_instance))
     }
 }