+++ src/level.rs
@@ -40,19 +40,19 @@ enum BackgroundImageError {
 
 fn background_image_sprite_sheet_bundle(
     images: &Assets<Image>,
-    texture_atlases: &mut Assets<TextureAtlas>,
+    texture_atlases: &mut Assets<TextureAtlasLayout>,
     background_image_handle: &Handle<Image>,
     background_position: &LevelBackgroundPosition,
     level_height: i32,
     transform_z: f32,
-) -> Result<SpriteSheetBundle, BackgroundImageError> {
+) -> Result<LdtkSpriteSheetBundle, BackgroundImageError> {
     if let Some(background_image) = images.get(background_image_handle) {
         // We need to use a texture atlas to apply the correct crop to the image
-        let tile_size = Vec2::new(
-            background_image.texture_descriptor.size.width as f32,
-            background_image.texture_descriptor.size.height as f32,
+        let tile_size = UVec2::new(
+            background_image.texture_descriptor.size.width,
+            background_image.texture_descriptor.size.height,
         );
-        let mut texture_atlas = TextureAtlas::new_empty(background_image_handle.clone(), tile_size);
+        let mut texture_atlas_layout = TextureAtlasLayout::new_empty(tile_size);
 
         let min = Vec2::new(
             background_position.crop_rect[0],
@@ -68,9 +68,7 @@ fn background_image_sprite_sheet_bundle(
 
         let crop_rect = Rect { min, max };
 
-        texture_atlas.textures.push(crop_rect);
-
-        let texture_atlas_handle = texture_atlases.add(texture_atlas);
+        let index = texture_atlas_layout.add_texture(crop_rect.as_urect());
 
         let scale = background_position.scale;
 
@@ -82,11 +80,17 @@ fn background_image_sprite_sheet_bundle(
         let center_translation =
             top_left_translation + (Vec2::new(scaled_size.x, -scaled_size.y) / 2.);
 
-        Ok(SpriteSheetBundle {
-            texture_atlas: texture_atlas_handle,
-            transform: Transform::from_translation(center_translation.extend(transform_z))
-                .with_scale(scale.extend(1.)),
-            ..Default::default()
+        Ok(LdtkSpriteSheetBundle {
+            sprite_bundle: SpriteBundle {
+                texture: background_image_handle.clone(),
+                transform: Transform::from_translation(center_translation.extend(transform_z))
+                    .with_scale(scale.extend(1.)),
+                ..Default::default()
+            },
+            texture_atlas: TextureAtlas {
+                index,
+                layout: texture_atlases.add(texture_atlas_layout),
+            },
         })
     } else {
         Err(BackgroundImageError::ImageNotLoaded)
@@ -210,7 +214,7 @@ pub fn spawn_level(
     commands: &mut Commands,
     asset_server: &AssetServer,
     images: &Assets<Image>,
-    texture_atlases: &mut Assets<TextureAtlas>,
+    texture_atlases: &mut Assets<TextureAtlasLayout>,
     ldtk_entity_map: &LdtkEntityMap,
     ldtk_int_cell_map: &LdtkIntCellMap,
     entity_definition_map: &HashMap<i32, &EntityDefinition>,